Refactor domain model and add value objects
Removed the empty `Class1` class from the `DomainDrivenDesignInsurance.Domain` namespace as it was no longer needed.

Added the `IAggregateRoot` interface to serve as a marker for aggregate root entities in the domain model.

Introduced the `Money` value object in the `DomainDrivenDesignInsurance.Domain.ValueObject` namespace to represent monetary values with properties for `Amount` and `Currency`. Included methods for addition, subtraction, and currency validation, along with equality checks and string representation.

Added the `PersonName` value object in the `DomainDrivenDesignInsurance.Domain.ValueObject` namespace to represent a person's name with properties for `FirstName` and `LastName`. Ensured proper validation, equality checks, and string representation.